requirements.txt
torch
transformers
bitsandbytes
accelerate
pillow
pypdf
olmocr
sentencepiece
